[web] Fixed page flickering and reorganized js code.
