+++ .editorconfig
@@ -8,7 +8,7 @@ indent_size = 4
 [*.{md,pandoc}]
 trim_trailing_whitespace = false
 
-[*.{json,yml,toml,html,css}]
+[*.{json,yml,toml,html,css,js}]
 indent_size = 2
 
 [Jenkinsfile]
