+++ web/layouts/docs/single.html
@@ -27,7 +27,7 @@
   <div class="w-3/4">
     <h1 class="font-serif-caps">{{ .Title }}</h1>
     {{ if or (eq $subsection "devguide") (eq $subsection "quickstart") }}
-    <div class="mb-4" id="os-selector">Select OS:
+    <div class="invisible mb-4" id="os-selector">Select OS:
       <a class="btn" id="btn-win"><i class="fab fa-windows"></i> Windows</a>
       <a class="btn" id="btn-linux"><i class="fab fa-linux"></i> Linux</a>
       <a class="btn" id="btn-mac"><i class="fab fa-apple"></i> macOS</a>
